Update piku cli from github
Once time Piku is installed in a server there is not an way to update the cli. So using 'update' client we have a non secure, but usefull cli updater. The command downloads piku.py from master branch to a temp file, if http return code is 200, its enough to supose that downloaded file is valid. We dont have file signature to validate.